Always compress values when using a sink when exploding packs
Looks like IO is a major bottleneck, followed by gzip which just
does about 20MB per core.

Of course there is also cost of extracting the objects in the first
place, but the OUTPUT part seems particularly slow.

Git is a slow writer due to that, but if done right, I would think
100MB/s on my machine are possible when writing directly into packs.