Now I understand why there is a separate resolution phase…
…which we can bypass nonetheless.
Interesting is that in order to get Sha1's, we need to apply
deltas. That fortunately… is actually not that easy as we
have to traverse the delta tree properly, while keeping intermediates.

The trick could be to keep track of a recursive ref-count helping
us to know when we can stop keeping intermediate objects.

Also means we have to traverse the whole pack once while keeping
crucial information, to finally compute all hashes.

Definitely interesting, and yet another algorithm to try to
resolve objects efficiently.
Now it's the question how well that can be threaded, and how much
memory it will take at worst/best.