Simplify indexing step
That way, most of the work can (and should) be done during the resolving
phase, as this can be multithreaded.
Furthermore, one should try not to do any work really in the indexing
phase to not slow down the producer.

However, I feel that this can't really happen with hashes, but anyway,
we only do CRCs when it's easy to avoid having to cache more or deal
with even more complexity.