[clone] Handle remote progress name prefixing (more) correctly
Otherwise there would only be long columns of similar text.
It's certainly not coming in in small chunks.

What seems to be happening is that we read data but at some point
get zero sized output as there is only progress, and not a single usable
byte. Even though that doesn't really match the observations as it,
at once, decodes all progress.
Then it stops reading before all data has been transferred.