First pack-to-pack copy and crc32 verification
Even though it looks wasteful, we are only interested in base objects.
These are decompressed directly which is where we learn the length
of the compressed entry.

The latter we needed to take the compressed entry, crc check it and
move it forward to the destination.

In case of delta objects we have to put in the work to decode the
whole object and then re-encode it. Deltified objects can't
be used for pack to pack copy, at least not in the moment,
as it would require us to assure that all required bases
are actually available on the other end.

Maybe this can happen once 'thin-pack' is possible, even though
thinking about it seems difficult right now.