[git-transport] ARGH: PIN!!!
This is what happens if you forget that Pins are needed to call
AsyncWrite methods - these are only implemented if the receiver is
Pin. Makes sense, very much known, but stupid me sometimes forgets
and the compiler in this case really wasn't helpful.

Maybe one day it can propose trat a trait method requires a different
receiver.