[refs] allow writing any valid ref value instead of limiting ourselves to object ids
This is more consistent as an interface for changing values and allows
HEAD references to be detached and reattched using a single mechanism
which is more likely to get all the looking right.

Especially when ref-table shows up there should be no special handling
of HEAD as compared to other references.