[smart-release #162] FAIL: cannot use extension traits…
…as this makes borrowchk even more fuzzy - it looses knowledge about
what actually borrows from self, which is probably fair enough.

This means we can't easily or at all inherit implementations using this
way.

Next issue is that it seems impossible to keep a ref to self around
in the produced object, and…that makes sense since each object
would want access to it.