Address FIXME related to git_pack::data::Object (#250)
It's correct to have it in git-pack due to the pack-location field.
The latter is required for additional metadata to be available when
creating packs.

One might be able to re-arrange parts of the implementation so that
the non-packlocation fields live in git-object. However, the latter
wouldn't be very useful as of now as the `Find` trait does have to
return objects with pack location information.