Sketch types across git-features, git-odb and git-repository… (#259)
The goal is to make thread-safety toggalble with features to get type
complexity down to just a single type in git-repository without needing
more feature toggles in parent crates.

This should work actually, this is an intermediate commit.

Something that will change is the settings of Easy, as there probably
will only be an Easy and EasyShared, don't know how the types
can be filled in based on a feature toggle though.
Maybe it can work with more typedefs.