improve pack staleness handling (#259)
Since packs can only be loaded after loading indices, one will have a
marker that has to be used to check for changes that can't be
reconciled.

Note that this is not to protect from querying packs that aren't
available on disk anymore as we will never unload them anyway when
in a mode that needs pack-ids to remain stable.