change!: parallel utilities now use `Send + Clone` insted of `Send + Sync` (#263)
This helps to assure that thread-local computations always work with the
kind of types we provide. The ones that are carrying out actions are
notably not `Sync` anymore.

We cater to that by defining our bounds accordingly, but for those
who want to use other utilities that need Sync, using types like
`Repository` and `thread_local!()` is the only way to make this
work.