Improve diffing example to compare performance, right now it looks like a regression (#266)
Could be due to the new implementation going through a refcell, whereas
the previous one didn't.

That's fair, I guess, it gets quite a few calls, but also means that
counting with tree-diff got about 10% slower now that the cache is
built-in, maybe.

For max-speed, one would probably have a local cache to avoid
unnecessary object lookups for previously obtained information
alltogether.