Try to handle state differently, but… (#266)
…this has many problems.

However, it slowly manifests that ideally we are able to handle a
certain amount of files, mutate many of them at the same time, while
being lock-free.

It's something we can't have though, so maybe just operate on a single
data structure but bank on caching it in the handles itself so that
super-fast and entirely lock-free access isn't even required.

The thing we can do is synchronizing access to the data structure,
loading files one by one instead of in parallel, and claiming that
this initial delay that's longer than it would have to be is acceptable
knowing that from thereon out it will be fast.

It will be fully lazy as well.