Uncover slightly disturbing races which make it hard to ever release/unload maps (#266)
This is because ref-counts don't matter that much if somebody is going
to ask for a pack by index while it's about to be garbage collected.

There is a chance that never becomes a problem as these packs
can live long enough to be available, but it's a possibility
that we cannot detect while the race is ongoing.