Support for 'sdir' extension (#293)
It's really only a marker and it doesn't seem to matter much
as not reading it won't matter for the entire git test suite.
Nor does it matter when it's not written, and I wonder if
a sparse checkout without a directory is even possible. In the latter
case one would have to remember sparseness differently as there is no
entry that could carry the 'sparse' file mode.