Add breaking test with conflicting file in index (#293)
To obtain it, run ./t/t6427-diff3-conflict-markers.sh on git df3c41adeb212432c53d93ce6ace5d5374dc6e11
and `cd t/trash*/rebase` to find it in `.git/index`.

It looks like the padding-skip is handled incorrectly, which instead has
to be calculated as if it was layed out in a C structure.