Don't interpolate on Path creation due to missing context (#331)
The lack of context would make %(prefix)/… paths generally fail
which makes the API prone to misuse.

This probably means that another API needs to be added to facilitate
obtaining paths while providing a way to pass the context, too.

For now having the `Path` type as intermediary seems sufficient,
but also might point towards other a more approach.