Make `Path::interpolate()` more useful by returning an actual `PathBuf` (#331)
That way users will want to call it or something like it in order to
make use of the `Path`, which otherwise is just a bunch of bytes which
aren't very useful at all in standard Rust.

Keeping the `Path` wrapper type is probably the right choice as it
matches the current API and allows to get paths explicitly which can
then be interpolated in just another step.