change!: remove `Store` from public API (#301)
It is unclear if ref-tables, which are the reason for it to exist in
the first place, will fit into the concept as they might not support
worktrees. It's entirely unclear how this works.

Maybe there can be a non-worktree version of the store with work-trees
only being supported by the file based ref database, and ref-tables
remaining a server-side feature.