Assure windows home path is compared in absolute terms (#426)
In 'git', the passed path for ownership check probably is absolute
for reasons, but that's not an assumption we can make in `gitoxide`.

Hence we resolve the path fully before comparing one path with another.
Without this, using `gitoxide` powered applications from the current
working dir set to the home (probably) wouldn't work.