Pin clap to v3.1 to avoid surprising deprecation warnings
I'd rather not have to deal with the new API because there is no problem
with the old one for me.

Once 4.0 becomes the default, the work has to be put in to probably
achieve the same, but I hope it's not more wordy than before.
I also hope that inconveniences will be figured out by the folks in the
mean time.

Related to https://github.com/clap-rs/clap/issues/3822