fix!: improve normalization; assure no extra copies are made on query. (#331)
We now return our own content, rather than the originals with their
lifetimes, meaning we bind lifetimes of returned values to our own
`File` instance. This allows them to be referenced more often, and
smarter normalization assures we don't copy in the simple cases
either.

More tests were added as well.
This is breaking as lifetime changes can cause distruptions, and
`values?_as()` was removed as well as it's somewhat duplicate
to higher-level APIs and it wasn't tested at all.