change!: remove local-time-support feature toggle. (#331)
We treat local time as default feature without a lot of fuzz, and
will eventually document that definitive support needs a compile
time switch in the compiler (`--cfg unsound_local_offset` or something).

One day it will perish. Failure is possible anyway and we will write
code to deal with it while minimizing the amount of system time
fetches when asking for the current local time.