change!: remove `Permissions::git_dir` field entirely.
It was meant to help dealing with bailing out if the git dir isn't
fully trusted, but the way this was done was over-engineered especially
since the read-only permission level wasn't implemented at all.

That function is now performed by a new flag, the `bail_on_untrusted`
which is off by default.