feat: Add `permissions::Config::git_binary` field (#450)
When true, default false, inject the git installation configuration file
if present at the cost of one `git config` invocation.

Note that we rely on the underlying `git-config` crate to not load
duplicate files.

We also currently lie about the scope which is actually unclear - have
seen 'unknown' or normal scopes like `system`.