get more clarity about `git ls-remote` and `git fetch` (#450)
Both differ as pattern on the command-line of ls-remote are just that,
without needing to be parsed as ref-spec. Thus many pattenrs there
seem to work when they should fail (if they were a ref-spec).

It's worth noting though that patterns which are source-only can
technically work just like a pattern, which would certainly make
them more flexible.