feat: forward line-diffing capabilities curtesy of the `similar` crate. (#470)
This is a first and maybe the last step towards providing diffing
functionality, and it seems like the right choice to keep this in
similar and contribute there as needed. All algorithms are well
described and thus shouldn't be git-specific per-se, and `similar`
is the best the community has to offer.