Try even harder to find the original index during recursive base object resolution. (#536)
Hard to understand, but the gist seems to be that looking up the base
onbject recursively changes the snapshot layout which makes the
follow-up lookup see a different index, which then won't have the
original id.

To circumvent, we will try to reuse it, but resort to a full search
afterwards.

Unfortunately there is no way to reproduce this, it's just the very
special linux kernel pack I have here.