change!: remove `fs-jwalk-single-threaded` in favor of `fs-walkdir-parallel`.
This way, `jwalk` and the dependencies (and troubles) it brings have to
be opted in, but also allow other users to actually opt out while
allowing the `parallel` feature to be in effect.

In other words, previously the `parallel` feature conflated `jwalk`
dependencies into the tree, which isn't the case anymore.