feat: automatically handle `.keep` files after writing a pack bundle to disk.
The logic implemented here tries to do the right thing, that is when
we have reason to believe that the objects in the pack are linked up
with a ref, we delete the keep file automatically.

However, if there was no local ref edit as the ref specs didn't contain
local destinations, or if the pack was empty, then keep the .keep file
and leave it to the caller to handle.