respect the current 'is_sparse()` state when writing.
Previously it would have been possible to write the sparse extension
without any indication that it is actually needed.

For now we rely on a flag which is technically a cache, after all
Mode::DIR indicates whether an index entry is sparse or not, and
with it the index itself. At least so I think.