keep track of http related configuration keys.
We also select the most important few for implementation.
The most time-consuming part would be to implement overrides correctly
and validate it against the baseline of git, it's unclear how git
can be queried or made to leak certain information. Maybe by overriding
DNS mappings to let it go against an unknown local IP/port?