FAIL: try to apply maybe-async in a place where it's probably not possible.
The goal is to re-use the existing tests, but right now they only
compile in async mode as the `maybe-async` crates needs
a feature to be set. Doing so is hard(er) if it's not already used
in the main crate, which we do not and will do our best to avoid.