[refactor] Deduplicate Tag and TagRef
To do so, implement a conversion from a `&Tag` to a `TagRef`.

This wasn't possible with the existing reprensentation of `Tag` as the
target was represented by an `ObjectId` which internally stored only 20
bytes, but the `TagRef` type used a `BStr` that represented the 40 bytes
`oid` representation of that object.

To fix that issue, store the target as a 40 bytes `BString` in `Tag`.

Alternative: The existing conversion from `TagRef` to `Tag` could have been
used instead, but that would have required allocating memory for the entire
data-structure which felt heavy handed for a simple refactoring.