fix!: `absolutize()` now takes a mandatory `current_dir()` parameter and returns `Option<path>`
Previously the function was willing to return an empty path despite it
being invalid. With the `current_dir` being required, this won't be the
case anymore and will yield logically consistent results in all cases.

This forces the caller to deal with the relative path being invalid
or crafted to produce some other path, maybe to bypass sanity checks.