add test that actually uses chdir like in the real world.
That way we don't have to rely on the parent gitoxide directory
to be present which is a premis that isn't always fulfilled,
for example when running tests from a tarball.

We also retire a similar test that relied on the presence of
the parent `gitoxide` repository, which isn't anything we can
assume.

More test-cases have been added to valid certain assumptions
once more.