fix!: `max-pure` now builds without any C build tooling due to lack of `openssl-sys`. (#602)
To make this work, we leave the `reqwest` configuration to downstream crates.
Note that this means downstream will have to select their TLS settings
themselves, so builds may fail after upgrade until this is done.

An example for a `reqwest` configuration can be found in the
`gitoxide` Cargo.toml in the root of the `gitoxide` repository.