fix: handle `file:///C:/foo/bar` urls correctly on windows, as paths now are `C:\\foo\bar`.
These paths are created when using the `url::Url::from_file_path()`
family of methods, which adds an extra slash at the beginning of a
windows path which makes it invalid there unless there is further
processing.

This is now applied by using `url` features, making this case work
specifically. Note that all other attributes are still the same
and `git-url` generally tries to keep paths in tact to be a hybrid
of type that can handle any file system paths as well as actual urls.