fix: `gix clone ssh://...` won't deadlock anymore.
For `cargo` specifically we now parse stderr to see if permission errors
occour. This links stderr and stdout and we have to pass information from
a supervisor thread that parses stderr to stdout and use the information to
return a custom io error in time.
Now the algorithm is adjusted to never be able to deadlock, as the problem
is inherently racy and somewhat hard to implement it properly especially without
a good test suite built-into `gitoxde` - there are no ssh servers one can easily
spin up cross-platform.