Break cyclical dev dependencies
This commit breaks cyclical dev-dependencies.
These cause problems with rust analyzer which makes gitoxide development
with RA basically impossible.
git-testools is mostly used as a utility we can simply use the
crates.io version of the respective gitoxide crates to circumvent that
problem. However, most other crates require matching crate versions.
Here the integration tests were simply moved to a subcrate in the same
folder by adding a Cargo.toml file in the tests folder.

The hex_to_id function in git-testtools is a specical case
because that function returns a git_hash::ObjectId instance.
This ObjectId would belong to the crates.io version of git-hash and can
not be used in tests. To work around that hex_to_id is moved into the
individual tests as it's a very small one-line function.