feat!: gate all signal handling behind the `signals` feature toggle. (#339)
This change also consolidates all signal handling into its own module called
`signal` to provide reusable handlers and as well as well as signal initialization.

Note that the functions to cleanup tempfiles don't interact with the signal registry,
hence they still can be called without the `signals` feature enabled.

Note that this change sneakily fixes a bug that could have caused a `write_all()`
on a tempfile that was removed by a signal to enter an infinite loop.