chore: remove `dashmap` in favor of own sharded concurrent hashmap.
This speeds up multi-threaded counting greatly, and despite it
using shared memory which makes it quite wasteful, it is possible to
outperform `git` with it if enough cores are thrown at the problem.

Single-threaded performance is still lacking though, ultimately it needs
caches to accelerate the counting stage to hopefully be competitive.