fix!: Upgrade `tempfile` to 3.4, fixing a security issue.
This is not really a breaking change, but a tactical major version increment
to give `cargo` an escape hatch and let it use the vulnerable version for a little
longer.

This means this `gix` up until v0.40 can be used to pull in `tempfile` 3.3, and
all future versions will use tempfile 3.4 (or 3.X) for that matter.