fix!: rename `WithSidebands::read_line()` to `*::read_line_to_string()`.
Also, we remove the existing blocking `read_line()` override as it's not
representable in the async implementation anyway, to unify both implementations
once again.

The new name also makes sure that we are not accidentally calling a built-in
implementation on the `BufRead` trait as it's clear our method has to be called
under a different name.