feat!: allow to ignore IO errors when reading configuration files. (#790)
Sometimes, IO errors can occour consistently when reading from certain
paths (on Windows, really), and it should be possible to not only see
which path that is but also to ignore them entirely.

Now IO errors contain the path of the failed configuration file.
There is also a new option flag to ignore errors (off by default).