fix: static linked list delta cache with memory cap (#851)
Previously, the 64 slot big LRU cache for pack deltas didn't
use any memory limit which could lead to memory exhaustion in the
face of untypical, large objects.

Now we add a generous default limit to do *better* in such situations.
It's worth noting though that that even without any cache, the working
set of buffers to do delta resolution takes considerable memory, despite
trying to keep it minimal.

Note that for bigger objects, the cache is now not used at all, which probably
leads to terrible performance as not even the base object can be cached.