try to get improved hitrate for delta-cache by using it more, and fail (#851)
This change will put more of the delta-chain into the cache
which possibly leads to increased chances of cache-hits if objects
aren't queried in random order, but in pack-offset order.

Note that in general, it tends to be faster to not use any cache at all.

This change was pruned back right away as the major difference to git,
which does it by storing every object of the chain in the cache, is that
we don't share the cache among threads. This leaves a much smaller per-thread
cache size which really is a problem if the objects are large.

So instead of slowing pack access down by trying it, with the default cache being
unsuitable as it would evict all the time due to memory overruns, we do nothing
here and rather improve the performance when dealing with pathological cases
during pack traversal.