feat: Add `fetch::Arguments::is_stateless()` to aid proper use of arguments.
When arguments are used, haves are reset every round in stateless protocols, while
everything else is repeated. However, this also means that previously confirmed common
commits aren't repeated unless this is specifically implemented by the user of `Arguments`.

That caller can now easily determine if negotiations have to be compensated for.

Please note that `Arguments` explicitly doesn't implement repeating of all prior arguments, which
would also repeat a lot of *in-vain* haves.