feat!: `gix_traverse::commit::Ancestors` now returns rich commit information.
As part of the iteration, it will decode part of the commit and read parents and possibly
the commit time if time-based sorting is desired. These values will now be made available
instead of just the `id` of the commit.

This saves the caller time to parse the commit again in case it needs similar information.