fix!: improve conversion API of `Pipeline::convert_to_git(index_object)` to be more consistent and practical.
`index_object` now receives the path to the entry as first argument, notably in a data-type that is fitting for
the purpose.
Additionally, implement `Clone` for `Pipeline`.
We also make sure that the input stream is turned into a buffer in case there is no driver, previously filtering
had nothing to work with if there was no driver.

Further:

* make setting the process context more intuitive.
* fix a bug that would cause output buffers to be too small, leading to memory violations.
* don't encode to the same target encoding if it is UTF-8 (just like git)