fix: don't let scripts run on (potentially) partially extracted archives and don't create archives on windows
This could cause unrelated failures which didn't help debugging at all.
Also improve error messages when trying to delete a stale/incorrect script
output directory fails.
This can easily happen on windows when symlinks are involved.

We also stop creating archives on windows as the archive-metadata can't be deleted for some
reason, which means it stays in the created directory and may cause script failures.
This seems acceptable as windows users simply won't be able to create their own archives for accelration,
but could probably use WSL for it.