remove the `Format` type in favor of an incredibly simple streaming format
The idea is to only get the vital information in a format that can easily
be decoded by the API user. This way, all container formats are implemented
in the calling crate or application.

That way we avoid having to deal with feature toggles for the various utility
crates that would implement the stream formats.