remove hardcoded target dir from justfile
The hardcoded `./target` path fails for environments where Cargo is
configured to use a different target directory (for example via
`CARGO_TARGET_DIR`).

Since we are now passing absolute paths to the `journey.sh` scripts the
previous behavior of manipulating the path needed to be removed. Because
some personal scripts may have depended on that functionality, I opted
to conditionally apply the manipulations if the given paths are found to
be relative.