add note go ODB initialization in preparation for dealing with many small packs.
When fetching, currently each of these creates a possibly small pack.
We can tackle this in many areas, so ideally there is a best algorithm
for this, maybe it is indeed just recency, or there is a selection.
Have to figure that out.

Further, small packs should be exploded, using `fetch.unpackLimit`.