Fix assertion in gix-url fixture test
Because `diag_url` contains the raw protocol string the assertion failed
for `ssh+git`/`git+ssh` urls. The problem is that `scheme.as_str()`
returns just `ssh` in those cases. By switching the conversion we can
circumvent this behavior.

Forget the above, I am stupid (currently amending this commit). The
problem is that we do not parse `ssh+git`/`git+ssh` as SSH urls - there
was nothing wrong with the previous version. Git uses an enum internally
for different schemes and converts it to a string for the diagnose
output. This is just one case where our parsing differs to the one from
Git. Nevertheless, I keep the change, because it made the problem way
clearer.