Fix prints of panics in gix-url baseline test
Due to [known-limitations][1] `libtest_mimic` can not capture the output
of the tests it executes, which causes all panic messages from the asserts
to be printed twice: during execution and in the test summary. Note that
`cargo nextest` does not have this problem, because it captures output
itself.

The prints during execution are annoying because they brake the normal
formatting of `cargo test` and because multiple tests run in parallel
their output is mixed too. By setting a no-op panic hook we can avoid
those prints.

[1]: https://docs.rs/libtest-mimic/latest/libtest_mimic/#known-limitations-and-differences-to-the-official-test-harness