Rewrite gix-url parsing
This change is intentionally kept very simple and does not try to
immediately fix all now failing tests. When comparing the test output of
this implementation with the previous there are a few things worth
mentioning:

- all previously passing baseline tests still do so
- the new implementation correctly classifies paths starting with a `/`
  or `./` but containing a `:` as file paths
- the new implementation correctly classifies paths containing just
  letters before the first `:` as scp paths
- the new implementation fails all tests that check the returned errors
  for invalid paths
- the new implementation seems to fail for some windows paths

The next step is to evaluate the now failing tests, and either remove/rewrite
them if they are based on flawed assumptions (looking at you
'ssh_alias_needs_username_to_not_be_considered_a_filepath') or fix the new
implementation to make them pass again.

We can then work on making the baseline tests pass.

The change also keeps the previous parsing logic around and just removes
it from the module graph, to allow for local comparisons of the two
implementations during development.