fix: `gix-index` prefix matching should now work correctly with conflicting files.
It was done in a rush and lacks a lot of tests. At least now it
has a greater chance of working, as tests that would truly validate
this are still missing for a lack of test date. It can be produced
with `git update-index`, but it wasn't yet worth it.