chore: add more test cases to pin behaviour around string and path handling.
This is important to know how to consistently handle empty strings - should they
not be present? Or should they be flagged as error?

The answer here is that the caller has to deal with it, as we report these values
if they are present which only requires an equals sign.